% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.chart.CumReturns.R
\name{new.chart.CumReturns}
\alias{new.chart.CumReturns}
\title{A New chart.CumReturns Function}
\usage{
new.chart.CumReturns(R, wealth.index = FALSE, geometric = TRUE,
  legend.loc = NULL, colorset = (1:12), begin = c("first", "axis"), ...)
}
\arguments{
\item{R}{an xts, vector, matrix, data frame, timeSeries or zoo object of
asset returns}

\item{wealth.index}{if \code{wealth.index} is \code{TRUE}, shows the "value
of $1", starting the cumulation of returns at 1 rather than zero}

\item{geometric}{utilize geometric chaining (TRUE) or simple/arithmetic chaining (FALSE) to aggregate returns,
default TRUE}

\item{legend.loc}{places a legend into one of nine locations on the chart:
bottomright, bottom, bottomleft, left, topleft, top, topright, right, or
center.}

\item{colorset}{color palette to use, set by default to rational choices}

\item{begin}{Align shorter series to: \itemize{ \item first - prior value of
the first column given for the reference or longer series or, \item axis -
the initial value (1 or zero) of the axis.  }}

\item{\dots}{any other passthru parameters}
}
\description{
This function allows you to plot periodic returns with cumulation and is based on xts::plot.xts rather than base graphics
}
\examples{
library(xtsPerAnl)
library(PerformanceAnalytics)
library(xtsExtra)  #library(xts)
library(RColorBrewer)
data(managers)
new.chart.CumReturns(managers)
}
\keyword{new.chart.CumReturns}

