% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.chart.Drawdown.R
\name{new.chart.Drawdown}
\alias{new.chart.Drawdown}
\title{A New chart.Drawdown Function}
\usage{
new.chart.Drawdown(R, geometric = TRUE, legend.loc = NULL,
  colorset = (1:12), ...)
}
\arguments{
\item{R}{an xts, vector, matrix, data frame, timeSeries or zoo object of
asset returns}

\item{geometric}{utilize geometric chaining (TRUE) or simple/arithmetic chaining (FALSE) to aggregate returns,
default TRUE}

\item{legend.loc}{places a legend into one of nine locations on the chart:
bottomright, bottom, bottomleft, left, topleft, top, topright, right, or
center.}

\item{colorset}{color palette to use, set by default to rational choices}

\item{\dots}{any other passthru parameters}
}
\description{
This function allows you to plot drawdowns and is based on xts::plot.xts rather than base graphics
}
\examples{
library(xtsPerAnl)
library(PerformanceAnalytics)
library(xtsExtra)  #library(xts)
library(RColorBrewer)
data(managers)
new.chart.Drawdowns(managers)
}
\keyword{new.chart.Drawdown}

