% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.chart.RollingPerformance.R
\name{new.chart.RollingPerformance}
\alias{new.chart.RollingPerformance}
\title{A New chart.RollingPerformance Function}
\usage{
new.chart.RollingPerformance(R, width = 12, FUN = "Return.annualized", ...,
  ylim = NULL, main = NULL, fill = NA)
}
\arguments{
\item{R}{an xts, vector, matrix, data frame, timeSeries or zoo object of
asset returns}

\item{width}{number of periods to apply rolling function window over}

\item{FUN}{any function that can be evaluated using a single set of returns}

\item{ylim}{set the y-axis limit, same as in \code{\link{plot}}}

\item{main}{set the chart title, same as in \code{\link{plot}}}

\item{fill}{a three-component vector or list (recycled otherwise) providing 
filling values at the left/within/to the right of the data range. See the 
fill argument of \code{\link{na.fill}} for details.}

\item{\dots}{any other passthru parameters to \code{\link{plot}} or the
function specified}
}
\description{
This function allows you to plot rolling performance metrics and is based on xts::plot.xts rather than base graphics
}
\examples{
library(xtsPerAnl)
library(PerformanceAnalytics)
library(xtsExtra)  #library(xts)
library(RColorBrewer)
data(managers)
new.chart.RollingPerformance(managers)
}
\keyword{new.chart.RollingPerformance}

