% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.chart.TimeSeries.R
\name{new.chart.TimeSeries}
\alias{new.chart.TimeSeries}
\title{A New chart.Timeseries Function}
\usage{
new.chart.TimeSeries(R, auto.grid = TRUE, yaxis.right = FALSE, type = "l",
  lty = 1, lwd = 2, las = par("las"), main = NULL, ylab = NULL,
  date.format.in = "\%Y-\%m-\%d", date.format = NULL, xlim = NULL,
  ylim = NULL, element.color = "darkgray", colorset = (1:12),
  pch = (1:12), legend.loc = NULL, ylog = FALSE, cex.axis = 0.8,
  major.ticks = "auto", grid.color = "lightgray", grid.lty = "dotted",
  xaxis.labels = NULL, ...)
}
\arguments{
\item{R}{an xts, vector, matrix, data frame, timeSeries or zoo object of}

\item{auto.grid}{if true, draws a grid aligned with the points on the x and
y axes}

\item{yaxis.right}{if true, draws the y axis on the right-hand side of the
plot}

\item{type}{set the chart type, same as in \code{\link{plot}}}

\item{lty}{set the line type, same as in \code{\link{plot}}}

\item{lwd}{set the line width, same as in \code{\link{plot}}}

\item{las}{set the axis label rotation, same as in \code{\link{plot}}}

\item{main}{set the chart title, same as in \code{\link{plot}}}

\item{ylab}{set the y-axis label, same as in \code{\link{plot}}}

\item{date.format.in}{allows specification of other date formats in the data
object, defaults to "\%Y-\%m-\%d"}

\item{date.format}{re-format the dates for the xaxis}

\item{xlim}{set the x-axis limit, same as in \code{\link{plot}}}

\item{ylim}{set the y-axis limit, same as in \code{\link{plot}}}

\item{element.color}{provides the color for drawing chart elements, such as
the box lines, axis lines, etc. Default is "darkgray"}

\item{colorset}{color palette to use, set by default to rational choices}

\item{pch}{symbols to use, see also \code{\link{plot}}}

\item{legend.loc}{places a legend into one of nine locations on the chart:
bottomright, bottom, bottomleft, left, topleft, top, topright, right, or
center.}

\item{ylog}{TRUE/FALSE set the y-axis to logarithmic scale, similar to
\code{\link{plot}}, default FALSE}

\item{cex.axis}{The magnification to be used for axis annotation relative to
the current setting of 'cex', same as in \code{\link{plot}}.}

\item{major.ticks}{Should major tickmarks be drawn and labeled, default
'auto'}

\item{grid.color}{sets the color for the reference grid}

\item{grid.lty}{defines the line type for the grid}

\item{xaxis.labels}{Allows for non-date labeling of date axes, default is
NULL}

\item{\dots}{any other passthru parameters}
}
\description{
This function allows you to plot timeseries and is based on xts::plot.xts rather than base graphics.
}
\examples{
library(xtsPerAnl)
library(PerformanceAnalytics)
library(xtsExtra)  #library(xts)
library(RColorBrewer)
data(managers)
new.chart.TimeSeries(managers)
}
\keyword{new.chart.TimeSeries}

